requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
markdownify==0.12.1
pyyaml==6.0.1
pydantic==2.8.2
jinja2==3.1.4

# Optional: LLM integration using litellm
litellm==1.41.0

# Optional: RAG DB
langchain==0.2.14
langchain-community==0.2.12
sentence-transformers==3.0.1
chromadb==0.5.5
faiss-cpu==1.8.0
